python-app/pyproject.toml
[project]
name = "python-app"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiodns>=3.5.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "beautifulsoup4>=4.13.4",
    "prettytable>=3.16.0",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "sqlmodel>=0.0.24",
]

